V0.3.5 release
This appears to be a fairly stable version.
It correctly processes and displays both 3D and 2D point cloud packets.
The workmode settings uses the same undocumented set work mode packet command that the Unitree archive library uses to set the workmode.  Changing workmodes between 3D to 2D seems to have some firmware issues.  There may be a sequence of the workmode commands along with reset commands to get the unit to successfully be in 2D scan mode.  Once successfully in the mode the scan head is not always in the same rotation position.  I have not found a way to set that other physically restrict the head rotation (which I am reluctant to do).
This release has a stable L2lidar class, a stable point cloud window class.  Everything is ui.
Please note that there are minor changes to the unitree_lidar_protocols.h and unitree_lidar_utilities.h

+++ src/Settings.cpp
@@ -252,7 +252,7 @@ void MainWindow::loadSettings(bool resetRequested)
         defaultPCsettings.Distance =settings.value("Distance", 10.0).toDouble();
         defaultPCsettings.Yaw =settings.value("Yaw", 145.0).toDouble();
         defaultPCsettings.Pitch = settings.value("Pitch", 20.0).toDouble();
-        defaultPCsettings.PointSize = settings.value("PointSize", 1).toInt();
+        defaultPCsettings.PointSize = settings.value("PointSize", 1.0).toDouble();
         defaultPCsettings.MinDistance = settings.value("MinDistance",0.1).toDouble();
         defaultPCsettings.MaxDistance = settings.value("MaxDistance",10.0).toDouble();
